feat(BoardInitializer) : 불필요한 Initializer 삭제
