--- src/main/java/chess/domain/board/BoardInitializer.java
@@ -1,5 +1,0 @@
-package chess.domain.board;
-
-public interface BoardInitializer {
-    ChessBoard initialize();
-}
